refactor(contracts): extract getPauserSetContract helper and clean up env examples
