+++ gateway-contracts/.env.example
@@ -142,11 +142,6 @@ PAUSER_ADDRESS_0="0x6591319B97979Acc59b7191A8B4Ec381375bFc92" # accounts[23] (ad
 # Pauser 2
 PAUSER_ADDRESS_1="0xb19e21437c47A541842bB84b018d3955462B35De" # accounts[24] (address)
 
-# Pauser swap inputs
-# These are used by the pauser management tasks when swapping pausers.
-OLD_PAUSER_ADDRESS_0="0x6591319B97979Acc59b7191A8B4Ec381375bFc92" # accounts[23] (address)
-NEW_PAUSER_ADDRESS_0="0xb19e21437c47A541842bB84b018d3955462B35De" # accounts[24] (address)
-
 # Protocol payment prices
 # The prices are in $ZAMA base unit (using 18 decimals)
 INPUT_VERIFICATION_PRICE="10000000000000000000" # (uint256, 10 $ZAMA)
