+++ host-contracts/.env.example
@@ -30,11 +30,6 @@ PAUSER_ADDRESS_0="0xa44366bAA26296c1409AD1e284264212029F02f1" # accounts[2] (add
 # Pauser 2
 PAUSER_ADDRESS_1="0xc1d91b49A1B3D1324E93F86778C44a03f1063f1b" # accounts[3] (address)
 
-# Pauser swap inputs
-# These are used by the pauser management tasks when swapping pausers.
-OLD_PAUSER_ADDRESS_0="0xa44366bAA26296c1409AD1e284264212029F02f1" # accounts[2] (address)
-NEW_PAUSER_ADDRESS_0="0xc1d91b49A1B3D1324E93F86778C44a03f1063f1b" # accounts[3] (address)
-
 # Host contract address
 # These addresses are required during the production deployment, at which point these environment 
 # variables are dynamically set after contracts are deployed.
