+++ library-solidity/.env.example
@@ -17,8 +17,6 @@ COPROCESSOR_SIGNER_ADDRESS_2="0xCFbF539CB91c92ace0343c5B0487149Ad0b82078" # acco
 COPROCESSOR_SIGNER_ADDRESS_3="0xA951F315d5FD35Cac111dFB5250DF231FB8eF905" # account[14]
 NUM_PAUSERS="1"
 PAUSER_ADDRESS_0="0xa44366bAA26296c1409AD1e284264212029F02f1" # accounts[2] (address)
-OLD_PAUSER_ADDRESS_0="0xa44366bAA26296c1409AD1e284264212029F02f1" # accounts[2] (address)
-NEW_PAUSER_ADDRESS_0="0xc1d91b49A1B3D1324E93F86778C44a03f1063f1b" # accounts[3] (address)
 ETHERSCAN_API_KEY="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
 RPC_URL="http://127.0.0.1:8756"
 HARDHAT_NETWORK="hardhat"
