+++ library-solidity/tasks/utils/loadVariables.ts
@@ -1,5 +1,6 @@
 import dotenv from 'dotenv';
 import fs from 'fs';
+import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import path from 'path';
 
 const ADDRESSES_DIR = path.resolve(__dirname, '../../fhevmTemp/addresses');
@@ -30,4 +31,18 @@ export function loadAddressEnvVarsFromFile(fileName: string) {
 
 export function loadHostAddresses() {
   loadAddressEnvVarsFromFile(HOST_ADDRESSES_ENV_FILE_NAME);
+}
+
+export async function getPauserSetContract(useInternalProxyAddress: boolean, hre: HardhatRuntimeEnvironment) {
+  await hre.run('compile:specific', { contract: 'fhevmTemp/contracts/contracts/immutable' });
+
+  const deployerPrivateKey = getRequiredEnvVar('DEPLOYER_PRIVATE_KEY');
+  const deployer = new hre.ethers.Wallet(deployerPrivateKey).connect(hre.ethers.provider);
+
+  if (useInternalProxyAddress) {
+    loadHostAddresses();
+  }
+  const pauserSetAddress = getRequiredEnvVar('PAUSER_SET_CONTRACT_ADDRESS');
+
+  return hre.ethers.getContractAt('PauserSet', pauserSetAddress, deployer);
 }
